Simplify filter bar, small refactor for vehicle settings.
